+++ lib/rumbleship-beeline.d.ts
@@ -37,6 +37,15 @@ export declare class RumbleshipBeeline {
      */
     linkToSpan(target: HoneycombSpan): void;
     withSpan<T>(metadataContext: object, fn: (span: HoneycombSpan) => T, rollupKey?: string): T;
+    /**
+     *
+     * @param this
+     * @param metadata_context
+     * @param fn
+     *
+     * @NOTE You 99.99% want the fn to be `async` and await its result before returning.
+     *  If you don't, the wrapped cb is finished outside of context and trace is lost.
+     */
     withAsyncSpan<T>(this: RumbleshipBeeline, metadata_context: object, fn: (span: HoneycombSpan) => Promise<T> | T): Promise<T>;
     withTrace<T>(metadataContext: object, fn: () => T, withTraceId?: string, withParentSpanId?: string, withDataset?: string): T;
     startTrace(span_data: object, traceId?: string, parentSpanId?: string, dataset?: string): HoneycombSpan;